vbdev_crypto.c: add a bogus assert for scan-build
Scan-build fails to realize that we have cleared the spdk_mempool_create
mock variable in crypto_ut.c:745 so it claims that g_session_mp and
g_mbuf_mp may be equal. No matter how many ways we explicitly set that
variable, scan-build can't seem to figure it out.
I am in the process of getting registered with the project so that I can
file these bug reports.

Change-Id: Id0ef196620c2984252f792a7e40162e10b989748
Signed-off-by: Seth Howell <seth.howell@intel.com>