dif: Change the numeric value of DIF type enums.
In the End-to-End Data Protection Type Settings (DPS) field in the
Identify Name Space Data Structure of NVMe,
type 1 is 001b, type 2 is 010b, and type 3 is 100b.

For type 3, conversion between 100b and 011b should be avoided.

This patch changes the each numeric value of DIF type enums to
follow NVMe specification.

Change-Id: I7f680ac6e14de7d12a64b147437254edb8800e31
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>