rte_vhost: introduce vhost transport operations structure
This is the first of a series of patches whose purpose is to add support
for the virtio-vhost-user transport. This is a vhost-user transport
implementation that is different from the default AF_UNIX transport. It
uses the virtio-vhost-user device which tunnels vhost-user protocol
messages over virtio. This lets guests act as vhost device backends for
other guests.

In order to add support for additional transports, it is necessary to
extract transport-specific code from the main vhost-user code.

This patch introduces struct vhost_transport_ops and associates each
device with a transport. This structure contains pointers to
transport-specific methods. Core vhost-user code calls into
vhost_transport_ops to perform transport-specific operations.

For now, this structure is empty. In the following patches, it will be
filled with the necessary fields in order to fulfill the vhost transport
abstraction.

Several more patches follow this one to complete the task of moving
AF_UNIX transport code out of core vhost-user code.

Change-Id: I35c06155ee1f124e2202e31eb66c76c901533f78
Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>