Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> > Spdk upstream already supports reconnection.
> 
> Yes we know that.
> 
> > The purpose of reconnection is to achieve IO recovery transmission, not abort these inflight IO.
> 
> Yes, this is the part I'm asking about. How you know that target recovered properly? Changpeng is working on SCSI Persistent reservation. How you know that reservation survived the crash? If you have volatile storage (like Malloc Bdev) it will not survive crash. How you deal with that if only "re-queue" the IO to storage that is already gone. In my mind any disconnection should be promoted to device reset by QEMU side instead pretending that nothing happen. This way guest kernel can know that queues need to reinitialized and IO re-queued.

The spdk may be disconnected. In the current version, whether the backend is a persistent device or a malloc device，it will resubmit IO that was not previously processed.

In test ,due to out of order IO，we found that this will lead to some IO hung. You can think of this patch as a bug fix. 

In our product environment, the spdk may accidentally disconnect. In order to avoid affecting some services,we expect that after the spdk is restarted, the service will not produce IO hung.

This patch does not do any re-queue for IO. It just uses a bitmap to record the inflight IO in queue.

Patch-set: 10
Reviewer: Lin Li <1015052@d5d70762-12d0-45a1-890d-524b12d3f735>