Update patch set 10
Patch Set 10:

> In test ,due to out of order IO，we found that this will lead to some IO hung. You can think of this patch as a bug fix. 

>> IO zawiesza się, ponieważ jądro gościa nie wie o rozłączeniu vhost...

Sorry for that, It was late for me :P. This is the translation:

IO hang because the guest kernel does not know about disconnecting vhost. I'm talking about producing the device reset event on diconnect (or reconnect this need to be investigated) to let the guest kernel know that IO can be safly aborted or requeued and virtio rings need to be reinitialized. After disconnecting it is safe.

> If you think QEMU implementation is better, why does upstearm resubmit the previous IO after restart? Does this implementation take into account the non-persistent memory problem?

Are we talking about some portion of code? Do you have some link. Maybe I need to educate my self more in this area.

> Can this patch be turned on as a configurable option for some persistent devices?

If there will be no other option I think we will have to. But first lets investigate other ways of deal with this issue.

I'm just curiouse: did you try reseting the virtio device after reconnect? Is this solving the IO hang issue?

Patch-set: 10