rte_vhost: add virtio pci framework
The virtio-vhost-user transport will need a virtio pci device driver
because the virtio-vhost-user device is a virtio PCI device.

Currently, there is a virtio pci driver under lib/virtio. However, it
cannot be used as it is.

For now, we will copy the virtio pci code from dpdk/drivers/net as a
temporary solution just to get this working. We will come back to this
later.

Change-Id: I146939b9687a1d609fc08377057a926101a035cc
Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>