vhost: abstract vring call mechanism as it is transport-specific
vring notification mechanism is transport-specific. Therefore, we add a
new transport-specific method called "vring_call()" that has to be
implemented by every vhost-user transport.

At present, the vring notifications are triggered by vhost dataplane
code in lib/vhost/vhost.c. But the implementation of the vring
notification mechanism has to be placed in the rte_vhost library. This
implies that a generic notification function has to be added in the
rte_vhost public API. This function will be called
rte_vhost_vring_call().  We add its function prototype in rte_vhost.h to
make it public.

rte_vhost_vring_call() takes a vring_idx as an argument to associate the
struct spdk_vhost_virtqueue instance with the relevant struct
vhost_virtqueue instance. We insert a new vring_idx field in struct
spdk_vhost_virtqueue to enable this association. This field is
initialized in start_device().

Finally, a stub for rte_vhost_vring_call() is added in the vhost unit
test file.

Change-Id: I4199fdc35d4f441f7c32fd94a54bafb1cfc0f95a
Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>