rte_vhost: introduce transport-specific cleanup_device method
This patch inserts a new transport-specific function called
"cleanup_device". This function frees any resources associated with a
virtio device. In case of the AF_UNIX transport, it is unecessary.

Transport-specific cleanup_device() is called from the generic
cleanup_device() in rte_vhost/vhost.c.

It will be used in the virtio-vhost-user transport that will be
introduced in later patches.

Change-Id: I1f8f5d67775a512a25f61cfb1ba8067c318f086f
Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>