rte_vhost: add virtio-vhost-user transport
This patch inserts the virtio-vhost-user transport to the rte_vhost
library. The virtio-vhost-user transport is a vhost-user transport
implementation like the AF_UNIX transport. It uses the virtio-vhost-user
device to tunnel the vhost-user protocol messages over virtio from the
vhost-user socket on the host to the slave guest. This allows a guest to
act as a vhost device backend for other guests.

The virtio-vhost-user transport enables running the SPDK vhost target
inside a dedicated Storage Appliance VM instead of host user space.

For more information on the virtio-vhost-user device, see
https://wiki.qemu.org/Features/VirtioVhostUser.

Notice that this patch inserts the virtio-vhost-user transport but it is
not usable yet. The new transport will be integrated with the core vhost
code in the following patches.

Change-Id: I5a1184ca29e4ce226866f181781d072518a061f1
Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>