rte_vhost: add RTE_VHOST_USER_VIRTIO_TRANSPORT flag
Extend the <rte_vhost.h> API to support the virtio-vhost-user transport
as an alternative to the AF_UNIX transport. The caller provides a PCI
DomBDF address:

  rte_vhost_driver_register("0000:00:04.0",
                            RTE_VHOST_USER_VIRTIO_TRANSPORT);

The virtio-vhost-user transport continues to be unusable because we have
to extend the core vhost API as well. This will be done in the following
patches.

Change-Id: Ibff28a494eb179a5f16e756682c79a84b7dd709d
Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>