vhost: insert transport option in the vhost library
This patch inserts a transport option in the vhost library. This way,
the vhost library will be able to make use of the extended DPDK's
librte_vhost API, which supports the AF_UNIX and virtio-vhost-user
transports.

The chosen vhost-user transport is represented by a global variable in
`lib/vhost/vhost.c`. Its value is set by `spdk_vhost_set_transport()`.
This function is not currently being used but it will be used by the
vhost app in the following patches. The transport variable is being
passed as a flag to the `rte_vhost_driver_register()`.

Change-Id: Ic730df7b8ff2e5157c465d6facc206cedf643d69
Signed-off-by: Nikos Dragazis <ndragazis@arrikto.com>