vhost: insert transport option in the vhost library
This patch inserts a transport option in the vhost library. This way the
vhost library will make use of the extended rte_vhost API that was
updated in previous patches in order to support the virtio-vhost-user
transport.

The chosen vhost-user transport is represented by a global variable in
lib/vhost/vhost.c.  Its value is set by the spdk_vhost_set_transport().
This function is not currently being used but it will be used by the
vhost app in the following patches. The transport variable is being
passed as a flag to the rte_vhost_driver_register().

Change-Id: Ic730df7b8ff2e5157c465d6facc206cedf643d69
Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>