vhost: keep the PCI address of the vvu device in the path field
In case of virtio-vhost-user transport, the PCI address of the
virtio-vhost-user device is saved in the path field of struct
spdk_vhost_dev. This required some minor modifications in the code in
order to handle both cases for the contents of the path field.

Also, the vhost device's name field has to be replaced by the path field
in case we have to search for the existence of a vhost device.
Therefore, function spdk_vhost_dev_find_by_path is introduced.

Change-Id: Ia8baf7cf32d240d854524017512cdd2d75f5aee3
Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>