vhost: add command line option to choose transport
Add "-T" option in the set of application-specific command line options
of the vhost app in order to choose between unix and vvu transports.
Default is unix if not specified. "-S" option is restructured to
represent either the directory for the unix sockets to be created, in
case of unix transport, or the PCI address in BDF format for the
virtio-vhost-user device, in case of vvu transport. Typing
"app/vhost/vhost -h" gives some detail about how to use these options.

Example usage:

$ app/vhost/vhost -S /var/tmp -T unix -m 0x1
$ app/vhost/vhost -S 0000:00:07.0 -T vvu -m 0x1

Change-Id: I73e327e98c8ef41e927d62b5eadcff318e2ef872
Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>