rte_vhost: change readiness condition for the vhost virtqueues
This patch changes the condition for the readiness of a vhost virtqueue
in vq_is_ready(). The check for invalid eventfds is removed in order to
support the virtio-vhost-user transport. In case of the
virtio-vhost-user transport, the vhost-user eventfds have to be invalid
as there is no point in passing eventfds from the host to the guest. The
notification mechanism for this transport is based on doorbells.

Change-Id: Icbd20dc1c26eb3c5f8ccc5ada170082fa24fe3d0
Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>