vhost/scsi: remove hotremoved scsi targets on device stop
In cases where initiator closes the connection as soon
as it receives a hotremove event, there is a possibility
that SPDK vhost will stop the device before finishing
up the asynchronous target hotremoval. The target would
be either hotremoved once the device is started again
(and it registers its management poller again) or it
could cause a potential memory leak if the Vhost-SCSI
device is removed. Even though the SCSI target itself
is always freed, the hotremoval completion callback is
only called from the device poller [1]. At least in our
RPC case, not calling that callback results in leaking
a context object and some json data.

We fix the above by calling all hotremove callbacks just
after stopping a device.

[1] process_removed_devs()

Change-Id: Ibfd773e1ab82b63643c57d7a9d37304e3007e38b
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>