env_dpdk: handle vvu devices with the spdk_virtio driver
Use the spdk_virtio driver for the vvu devices instead of the dedicated
virtio-vhost-user driver in trans_virtio_vhost_user.c. This makes sense
as the virtio-vhost-user device is a virtio device. This patch adds the
device ids for the virtio-vhost-user device to the spdk_virtio driver's
device id list.

This is the first step into integrating the vvu device with the
spdk_virtio driver.

Change-Id: I93b0ca8f688dc7c6073aabad93aeca5f5ec04b1f
Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>