vhost: attach vvu device to the spdk_virtio driver
This patch completes the transaction from the virtio-vhost-user driver
to the spdk_virtio driver.

The vvu device is attached to the spdk_virtio driver during vhost device
registration in spdk_vhost_dev_register(). This is done using function
spdk_pci_device_attach(). Function vvu_pci_probe() is passed as a
callback.

Using vvu_pci_probe() as a callback requires some changes in the
rte_vhost library. In specific, the function prototype of
vvu_pci_probe() has to be changed. This involves using spdk_pci_device
instead on rte_pci_device. Structure spdk_pci_device is defined in env.h
so this header file is included in rte_vhost.h. This triggers a conflict
between linux/if.h and net/if.h which is resolved by removing linux/if.h
from vhost.h. What is more, vvu_pci_probe() has to be added to the
public API of rte_vhost library in order to be visible in lib/vhost.c.
Therefore, its function prototype is added in rte_vhost.h

The vvu device is detached from the spdk_virtio driver when the vhost
device is removed. This makes sense because a vvu device can support
only one vhost controller. This means that a vvu device is tightly
related to one vhost controller. The detach is done using function
spdk_pci_device_detach() which is called in vvu_pci_remove() which in
turn is called in vvu_socket_cleanup(). It is possible to use
spdk_pci_device_detach() from within rte_vhost library because, as it is
mentioned above, the spdk/env.h has been included in rte_vhost.h.

Detaching the vvu device alongside with the vhost controller's removal
makes RPC "remove_vhost_controller" to work as expected in case of vvu
transport. This means that reusing the same vvu device to construct a
new vhost controller is now possible.

Last but not least, the necessary stubs are added in the vhost unit test
file.

Change-Id: Ide23632ec265bd7c9ec16cb338c1e11319ccd518
Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>