Update patch set 1
Patch Set 1:

Discovered a few things trying to look at this:
* crypto UT will fail without ASAN enabled - this is new starting with a change to address a scan build issue a while back.  Probably no big deal but FYI...
* I had to comment out building COMPRESSDEV in DPDK on this patch to get things to build??

As the patch is now, UT for crypto do not pass and not because of the issue above.  You can see the failure in CH TP Ubuntu machine.

I also discovered that crypto UT are not running at all on Jenkins.  Will wait to address that in the morning as there's no harm in going another day w/o them, seems clear why they are not but I want to make sure I'm not missing something before I fix it.

Patch-set: 1
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>