Update patch set 1
Patch Set 1:

checked the current code, we did not tell the Guest kernel that we support flush command. But when we test aio, if flush is not supported, the data will not be flushed to the device managed by AIO device. Changpeng will provide a patch to fix this. 
1 If the bdev is supporting flush I/O type, we tell the guest that we will support that.
2 Add some missing code to handle the flush command (e.g., in vhost blk module) to pass the command to the underlying bdev.

Patch-set: 1
Reviewer: Gerrit User 1011275 <1011275@d5d70762-12d0-45a1-890d-524b12d3f735>