[Discussion]blobfs: fix the length value of file.
In function spdk_fs_file_stat_async, the
stat.size = f->append_pos >= f->length ? f->append_pos : f->length;

but in spdk_file_get_length, we return f->length.

So I am confused, why we return different value. I know that
in the processsing, append_pos will be increased. So my question
is that the functin of stat.size and spdk_file_get_length should
be equal. This may be the reason that we found the size reported by
application and SPDK are not same.

Change-Id: Idb9aa9e737711fcd48ac0075c7f7ffed825fe3b0
Signed-off-by: Ziye Yang <optimistyzy@gmail.com>