nvme_perf: Enable SGL for I/O commands.
Most use cases of NVMe I/O commands have used SGL now, and
upcoming DIF/DIX support will use it too. Hence this patch adds
a global option to use SGL.

AIO also have supported SGL (IOCB_CMD_PREADV and IOCB_CMD_PWRITEV)
but it may not be available all environment and is not so critical
for SPDK. Hence enabling SGL for AIO is not included in this patch.

Change-Id: I2ae185f7819964d6125e26fe27f5d02ffa2cac2e
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>