Update patch set 2
Patch Set 2:

I'm not confident in my understanding, but this patch is to fix the bug? or investigate the cause of some observed issues?

If the purpose is the latter, how about adding the current subsystem state, subsystem->state to the error log? 

It's OK if you have the logic that result in the issue and don't open the window so widely.

Patch-set: 2
Reviewer: Gerrit User 1012251 <1012251@d5d70762-12d0-45a1-890d-524b12d3f735>