Update patch set 4
Patch Set 4: Code-Review-1

(3 comments)

Hey Liang,
Testing hotplug is going to be awesome! I think this is a really thorough description. I have a couple of questions though. Are we intending to put all of these tests in the per-patch test suite?
If so, do we need all of these redundant pathways for hotplug, or is there a minimum subset of them that can give us the same coverage of the nvme hotplug related code? This is just really important for timing when running per-patch testing.

Patch-set: 4
Reviewer: Gerrit User 1011223 <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1