nvme_perf: Use the new DIF library in Perf tool
Apply the new DIF library to IO submission and verification.

Condition to use DIF generation and verification are not changed
in this patch. Add an iovec entry to pass DIF generation and
verification. IO submission will use the iovec entry in the later
patch. Besides, this patch stops to clear metadata fields at read
IO submission. They will be overwritten and can be skipped.

Change-Id: I68ccfb41cce2546d7145417c1490f49f0e62bb7a
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>