Update patch set 2
Patch Set 2: Code-Review+2

This patch is fine - but reading the unit tests makes me think they should be adjusted to calculate the CRC dynamically instead of hardcoding the CRCs in the unit tests.  Otherwise if we get a failure, we don't know if it was because something like the my_lba didn't match, or because the CRCs were incorrect.

Patch-set: 2
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2