Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 6:
> > 
> > (2 comments)
> 
> Paul, I just pushed a vtophys_ut extension [1] that allocates a non-contiguous-buffer like the one we're trying to support here. Can I ask you to prepare a crypto example that uses such buffer?
> 
> [1] https://review.gerrithub.io/c/spdk/spdk/+/440531 test/vtophys: add a case with non-physically-contiguous buffer

So the problem with a UT like this on the crypto side is that it's a UT and there is no real crypto :) I assume that's what you were asking.  If we want to test this in a system test we'd need to inject something to create a split buffer back from spdk_vtophys().  But what I can try now is using your patch and manually (via gdb) forcing a split to while running bdevperf with verify on or something like that.  I'll give that a try

Patch-set: 6
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>