Update patch set 6
Patch Set 6:

I've tried multiple ways of dealing with this w/o chaining by manually forcing one 512B translation to come back as 256.  As explained above the results are 2 crypto operations instead of one and regardless of how I manipulate the IV and the offsets it doesn't work.  This is as I would expect based on how the IV was explained to me, once you've encrypted with a specific IV it must decrypted with the same length and same IV that was used at encryption.  That's why choosing LBA as IV works because we already read in multiples of LBA on disk but this makes a requirement on the crypto operations in memory that they also be LBA sized.

So, I believe the only way to address this is to support chaining for QAT which allows a single crypto operation (and IV) to be described my multiple mbufs.  Let me know your thoughts, I had chaining in the driver once a long time ago but using LBA as IV seemed to simplify it to the point of not being needed, oh well :)

I can add chaining support if you agree with my assessment.

Patch-set: 6
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>