Update patch set 2
Patch Set 2:

Hi Jim,

We didn't see any error for SPDK master branch with this patch after a night.

We saw temporary out of memory for our SPDK application with this patch because of shortage of session pool but that is not related with this patch and SPDK master branch.

Hence we can close the issue.

Patch-set: 2
Reviewer: Gerrit User 1012251 <1012251@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0