nvmf/tcp: fix the issues of tcp requests recycling.
For tcp requests are in TCP_REQUEST_STATE_TRANSFERRING_CONTROLLER_TO_HOST
are not totally freed, because they may still in queued_c2h_data_tcp_req list.
(Only part of data may be sent).

Currrently, We only cleanup the list, but we did not free those reqs yet.
So It is still need to call: spdk_nvmf_tcp_drain_state_queue on this state
again.

Moreover, if the tqpair is in EXITED state, we do not need to handle
the queued_c2h_data_tcp_req, to avoid the recursive invoke when invoke
pdu->cb_fn.

Change-Id: I5856c3421019ec49d576d3dae4c62fefbb3925ca
Signed-off-by: Ziye Yang <optimistyzy@gmail.com>