Update patch set 48
Patch Set 48:

> Patch Set 34: Code-Review-1
> 
> Hi, 
> 
> As I understand, this patch introduces multiple RDMA_READ/RDMA_WRITE for transfering multi-list SGL. Today, RDMA receive/send queues capacity is set based on assumption that RDMA_READ/RDMA_WRITE operation need per request. Ib believe, that should be changed as part of this work.
> 
> Increasing queues size will affect scalability. It would be nice to see some estimations or even test results for this feature.

Hi Sasha,

I have been thinking a lot about this, and while this patch does introduce multiple read and write operations for a single request, it does so only on an as needed basis for multi-element SGLs. By keeping track of the number of outstanding work requests, we can avoid increasing the send queue size. So the only added memory overhead is the shared group of work request data structures. This is constant per transport and doesn't necessarily have to increase as more qpairs are added. Now, for workloads where all of the requests are large and multi-SGL across multiple qpairs, having a large shared pool of Work requests would be desirable for optimal performance.
One thing that I was careful to do is move the allocation of buffers and work requests to an early stage of the request processing so that when we run into cases when there are too many requests and not enough shared buffers or work requests, we are able to bail early and wait for more resources. This ensures that the performance of the multi-sgl path beats that of sending each sgl element on it's own NVMe-oF request even in cases of high queue depth and buffer scarcity.

I have tested this with a single qpair with I/O sizes ranging from 8k - 128k, sgl element sizes from 4k - 64k, and queue depth from 4 - 128. I found the Multi-SGL path to be more performant than the legacy behavior (splitting each sgl element into its own request) in every case.

Patch-set: 48
Reviewer: Gerrit User 1011223 <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>